Tidied up the code, changed package, removed external deps, added a demo usage, it should be okay to start things with.
