+++ src/main/java/com/github/guikeller/jtide/models/Coefficient.java
@@ -1,0 +1,21 @@
+package com.github.guikeller.jtide.models;
+
+public class Coefficient {
+
+  private String name = "";
+  private double value = 0D;
+
+  public Coefficient(String name, double d) {
+    this.name = name;
+    this.value = d;
+  }
+
+  public String getName() {
+    return name;
+  }
+
+  public double getValue() {
+    return value;
+  }
+
+}
