+++ src/main/java/com/github/guikeller/jtide/models/Constituents.java
@@ -1,0 +1,14 @@
+package com.github.guikeller.jtide.models;
+
+import java.util.LinkedHashMap;
+import java.util.Map;
+
+public class Constituents {
+
+  private Map<String, ConstSpeed> constSpeedMap = new LinkedHashMap<>();
+  
+  public Map<String, ConstSpeed> getConstSpeedMap() {
+    return constSpeedMap;
+  }
+
+}
