+++ src/main/java/com/github/guikeller/jtide/services/DoneWithSiteException.java
@@ -1,0 +1,13 @@
+package com.github.guikeller.jtide.services;
+
+import org.xml.sax.SAXException;
+
+public class DoneWithSiteException extends SAXException {
+
+    public final static long serialVersionUID = 1L;
+
+    public DoneWithSiteException(String s) {
+        super(s);
+    }
+
+}
