--- src/main/java/es/tidetim/tideengine/models/Constituents.java
@@ -1,67 +1,0 @@
-package es.tidetim.tideengine.models;
-
-import java.io.Serializable;
-
-import java.util.HashMap;
-import java.util.LinkedHashMap;
-import java.util.Map;
-
-public class Constituents implements Serializable
-{
-  @SuppressWarnings("compatibility:9177290185319880922")
-  private final static long serialVersionUID = 1L;
-  private Map<String, ConstSpeed> constSpeedMap = new LinkedHashMap<String, ConstSpeed>();
-  
-  public Map<String, ConstSpeed> getConstSpeedMap()
-  {
-    return constSpeedMap;
-  }
-    
-  public static class ConstSpeed implements Serializable
-  {
-    @SuppressWarnings("compatibility:-5425660072213952784")
-    private final static long serialVersionUID = 1L;
-
-    private int idx = 0;
-    private String coeffName = "";
-    private double coeffValue = 0d;
-    private Map<Integer, Double> equilibrium = new HashMap<Integer, Double>();
-    private Map<Integer, Double> factors     = new HashMap<Integer, Double>();
-    
-    public ConstSpeed(int idx, String name, double val)
-    {
-      this.idx = idx;
-      this.coeffName = name;
-      this.coeffValue = val;
-    }
-    
-    public void putEquilibrium(int year, double val)
-    {
-      equilibrium.put(new Integer(year), new Double(val));
-    }
-    public void putFactor(int year, double val)
-    {
-      factors.put(new Integer(year), new Double(val));
-    }
-
-    public String getCoeffName()
-    {
-      return coeffName;
-    }
-
-    public double getCoeffValue()
-    {
-      return coeffValue;
-    }
-
-    public Map<Integer, Double> getEquilibrium()
-    {
-      return equilibrium;
-    }
-
-    public Map<Integer, Double> getFactors()
-    {
-      return factors;
-    }
-  }
-}
