--- src/main/java/es/tidetim/tideengine/models/TideType.java
@@ -1,16 +1,0 @@
-package es.tidetim.tideengine.models;
-
-
-public enum TideType {
-    HW("High Water"), LW("Low Water"), FALLING("Falling"), RISING("Rising");
-
-    private String description;
-
-    TideType(String description) {
-        this.description = description;
-    }
-
-    public String getDescription() {
-        return description;
-    }
-}
