--- src/main/java/es/tidetim/tideengine/util/JsonDateDeserializer.java
@@ -1,24 +1,0 @@
-package es.tidetim.tideengine.util;
-
-import com.fasterxml.jackson.core.JsonParser;
-import com.fasterxml.jackson.core.ObjectCodec;
-import com.fasterxml.jackson.databind.DeserializationContext;
-import com.fasterxml.jackson.databind.JsonDeserializer;
-import com.fasterxml.jackson.databind.node.TextNode;
-
-import java.io.IOException;
-import java.time.LocalDateTime;
-
-public class JsonDateDeserializer extends JsonDeserializer<LocalDateTime> {
-
-    @Override
-    public LocalDateTime deserialize(JsonParser jp, DeserializationContext ctxt)
-            throws IOException {
-
-        ObjectCodec oc = jp.getCodec();
-        TextNode node = oc.readTree(jp);
-        String dateString = node.textValue();
-
-        return LocalDateTime.parse(dateString, JsonDateSerializer.formatter);
-    }
-}
