--- src/main/java/es/tidetim/tideengine/util/JsonDateSerializer.java
@@ -1,22 +1,0 @@
-package es.tidetim.tideengine.util;
-
-import com.fasterxml.jackson.core.JsonGenerator;
-import com.fasterxml.jackson.databind.JsonSerializer;
-import com.fasterxml.jackson.databind.SerializerProvider;
-
-import java.io.IOException;
-import java.time.LocalDateTime;
-import java.time.format.DateTimeFormatter;
-
-public class JsonDateSerializer extends JsonSerializer<LocalDateTime> {
-
-    public static final DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm:ss");
-
-    @Override
-    public void serialize(LocalDateTime date, JsonGenerator generator,
-                          SerializerProvider provider) throws IOException {
-
-        String dateString = date.format(formatter);
-        generator.writeString(dateString);
-    }
-}
