--- src/main/resources/application.yml
@@ -1,14 +1,0 @@
-application:
-  name: tides
-
-elasticsearch:
-  cluster: mygovscot
-  host: localhost
-  port: 9300
-  indexName: tides
-  
-logging:
-    level:
-        es.tidetim: DEBUG
-        org.springframework: INFO
-
