+++ src/main/resources/xml/README.txt
@@ -1,0 +1,3 @@
+XMLs downloaded from the NOAA ( I reckon )
+Or perhaps from IHO ( TWCWG )
+Asked: https://github.com/crystalmark/tides/issues/2
